split controller.py into polling_thread.py
